+++ android/src/main/java/io/openim/flutter_openim_sdk/manager/ConversationManager.java
@@ -17,6 +17,12 @@ public class ConversationManager extends BaseManager {
         Open_im_sdk.getAllConversationList(new BaseListener(result));
     }
 
+    public void getConversationListSplit(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.getConversationListSplit(new BaseListener(result),
+                int2long(methodCall, "offset"),
+                int2long(methodCall, "count"));
+    }
+
     public void getOneConversation(MethodCall methodCall, MethodChannel.Result result) {
         Open_im_sdk.getOneConversation(
                 value(methodCall, "sourceID"),
@@ -70,5 +76,20 @@ public class ConversationManager extends BaseManager {
         CommonUtil.runMainThreadReturn(result, Open_im_sdk.getConversationIDBySessionType(
                 value(methodCall, "sourceID"),
                 int2long(methodCall, "sessionType")));
+    }
+
+    public void setConversationRecvMessageOpt(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.setConversationRecvMessageOpt(
+                new BaseListener(result),
+                jsonValue(methodCall, "conversationIDList"),
+                int2long(methodCall, "status")
+        );
+    }
+
+    public void getConversationRecvMessageOpt(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.getConversationRecvMessageOpt(
+                new BaseListener(result),
+                jsonValue(methodCall, "conversationIDList")
+        );
     }
 }