+++ ios/Classes/Module/ConversationManager.swift
@@ -15,6 +15,7 @@ public class ConversationManager: BaseServiceManager {
         
         self["setConversationListener"] = setConversationListener
         self["getAllConversationList"] = getAllConversationList
+        self["getConversationListSplit"] = getConversationListSplit
         self["getOneConversation"] = getOneConversation
         self["getMultipleConversation"] = getMultipleConversation
         self["deleteConversation"] = deleteConversation
@@ -24,6 +25,8 @@ public class ConversationManager: BaseServiceManager {
         self["markGroupMessageHasRead"] = markGroupMessageHasRead
         self["getTotalUnreadMsgCount"] = getTotalUnreadMsgCount
         self["getConversationIDBySessionType"] = getConversationIDBySessionType
+        self["setConversationRecvMessageOpt"] = setConversationRecvMessageOpt
+        self["getConversationRecvMessageOpt"] = getConversationRecvMessageOpt
     }
     
     func setConversationListener(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
@@ -34,7 +37,11 @@ public class ConversationManager: BaseServiceManager {
     func getAllConversationList(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
         Open_im_sdkGetAllConversationList(BaseCallback(result: result))
     }
-    
+
+    func getConversationListSplit(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
+        Open_im_sdkGetConversationListSplit(BaseCallback(result: result), methodCall[int: "offset"], methodCall[int: "count"])
+    }
+
     func getOneConversation(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
         Open_im_sdkGetOneConversation(methodCall[string: "sourceID"], methodCall[int: "sessionType"], BaseCallback(result: result))
     }
@@ -70,6 +77,14 @@ public class ConversationManager: BaseServiceManager {
     func getConversationIDBySessionType(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
         let conversationID = Open_im_sdkGetConversationIDBySessionType(methodCall[string: "sourceID"], methodCall[int: "sessionType"])
         callBack(result, conversationID)
+    }
+
+    func setConversationRecvMessageOpt(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
+        Open_im_sdkSetConversationRecvMessageOpt(BaseCallback(result: result), methodCall[jsonString: "conversationIDList"], methodCall[int: "status"])
+    }
+
+    func getConversationRecvMessageOpt(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
+        Open_im_sdkGetConversationRecvMessageOpt(BaseCallback(result: result), methodCall[jsonString: "conversationIDList"])
     }
 }